feat(dgeni): consolidate docs nav list generation (#2989)
