+++ tools/dgeni/src/templates/api/navigation-list.template.json
