+++ tools/dgeni/src/templates/guides/navigation-list.template.json
