+++ tools/dgeni/src/utils/capitalize.ts
