+++ tools/dgeni/src/utils/configurator/output.ts
@@ -5,25 +5,26 @@ import {
   DAFF_DOC_KIND_PATH_SEGMENT_MAP,
 } from '@daffodil/docs-utils';
 
-import { Configurator } from '../../../../utils/configurator.type';
-import { GenerateGuideListProcessor } from '../../processors/generateGuideList';
+import { Configurator } from './type';
+import { GenerateNavListProcessor } from '../../processors/generateNavList';
 
 export interface OutputPathsConfig {
   kind: DaffDocKind;
   outputPath: string;
+  docTypes: Array<string>;
 }
 
 export const outputPathsConfigurator: Configurator<OutputPathsConfig> = (config: OutputPathsConfig) => (pkg: Package) => pkg
-  .processor(new GenerateGuideListProcessor())
-  .config((generateGuideList: GenerateGuideListProcessor) => {
-    generateGuideList.outputFolder = `${config.outputPath}/${DAFF_DOC_KIND_PATH_SEGMENT_MAP[config.kind]}`;
+  .processor(new GenerateNavListProcessor())
+  .config((generateNavList: GenerateNavListProcessor) => {
+    generateNavList.outputFolder = `${config.outputPath}/${DAFF_DOC_KIND_PATH_SEGMENT_MAP[config.kind]}`;
   })
   .config((computePathsProcessor) => {
     computePathsProcessor.pathTemplates.push({
-      docTypes: ['guide'],
+      docTypes: config.docTypes,
       getPath: (doc) => {
         doc.moduleFolder = `${config.outputPath}/${DAFF_DOC_KIND_PATH_SEGMENT_MAP[config.kind]}/${doc.id}`;
-        return `/${doc.moduleFolder}`;
+        return doc.moduleFolder;
       },
       outputPathTemplate: '${moduleFolder}.json',
     });
