+++ tools/dgeni/src/utils/configurator/type.ts
