+++ tools/dgeni/src/utils/navigation-trie.ts
